Refactored PropertiesItem UserControl into a Templated Control
With a new PropertiesViewItem as a control, with other controls derived from it, we can share code and avoid duplicating Dependency Properties.

Started work on a Ratings variation of the control
