--- src/Files.App.Controls/PropertiesView/PropertiesItemText.xaml
@@ -1,117 +1,0 @@
-<!--  Copyright (c) Files Community. SPDX-License-Identifier: MPL-2.0  -->
-<UserControl
-	x:Class="Files.App.Controls.PropertiesItemText"
-	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-	mc:Ignorable="d">
-	<Grid
-		Padding="12,6"
-		HorizontalAlignment="Stretch"
-		VerticalAlignment="Stretch">
-		<Grid.ColumnDefinitions>
-			<ColumnDefinition Width="{x:Bind LabelWidth}" />
-			<ColumnDefinition Width="*" />
-		</Grid.ColumnDefinitions>
-		<Grid.RowDefinitions>
-			<RowDefinition Height="Auto" />
-			<RowDefinition Height="Auto" />
-		</Grid.RowDefinitions>
-
-		<TextBlock
-			x:Name="Label_PART"
-			Grid.Row="0"
-			Grid.RowSpan="2"
-			Grid.Column="0"
-			Grid.ColumnSpan="1"
-			Margin="0,2,12,2"
-			VerticalAlignment="Center"
-			Foreground="{ThemeResource TextFillColorSecondaryBrush}"
-			Style="{ThemeResource CaptionTextBlockStyle}"
-			Text="{x:Bind Label}" />
-
-		<TextBlock
-			x:Name="Text_PART"
-			Grid.Row="0"
-			Grid.RowSpan="2"
-			Grid.Column="1"
-			Grid.ColumnSpan="1"
-			Margin="0,2,12,2"
-			VerticalAlignment="Center"
-			Text="{x:Bind Text, Mode=TwoWay}"
-			TextWrapping="WrapWholeWords"
-			Visibility="Visible" />
-
-		<TextBox
-			x:Name="TextEdit_PART"
-			Grid.Row="0"
-			Grid.RowSpan="2"
-			Grid.Column="1"
-			Grid.ColumnSpan="1"
-			Margin="-12,2,0,2"
-			VerticalAlignment="Center"
-			Text="{x:Bind Text, Mode=TwoWay}"
-			Visibility="Collapsed" />
-
-
-
-
-		<VisualStateManager.VisualStateGroups>
-			<VisualStateGroup x:Name="EditableStates">
-				<VisualState x:Name="NotEditableState">
-					<VisualState.Setters>
-						<Setter Target="TextEdit_PART.Visibility" Value="Collapsed" />
-						<Setter Target="Text_PART.Visibility" Value="Visible" />
-					</VisualState.Setters>
-				</VisualState>
-				<VisualState x:Name="EditableState">
-					<VisualState.Setters>
-						<Setter Target="TextEdit_PART.Visibility" Value="Visible" />
-						<Setter Target="Text_PART.Visibility" Value="Collapsed" />
-					</VisualState.Setters>
-				</VisualState>
-			</VisualStateGroup>
-			<VisualStateGroup x:Name="OrientationStates">
-				<VisualState x:Name="HorizontalState">
-					<VisualState.Setters>
-						<Setter Target="Label_PART.(Grid.Row)" Value="0" />
-						<Setter Target="Label_PART.(Grid.RowSpan)" Value="2" />
-						<Setter Target="Label_PART.(Grid.Column)" Value="0" />
-						<Setter Target="Label_PART.(Grid.ColumnSpan)" Value="1" />
-
-						<Setter Target="TextEdit_PART.(Grid.Row)" Value="0" />
-						<Setter Target="TextEdit_PART.(Grid.RowSpan)" Value="2" />
-						<Setter Target="TextEdit_PART.(Grid.Column)" Value="1" />
-						<Setter Target="TextEdit_PART.(Grid.ColumnSpan)" Value="1" />
-
-						<Setter Target="Text_PART.(Grid.Row)" Value="0" />
-						<Setter Target="Text_PART.(Grid.RowSpan)" Value="2" />
-						<Setter Target="Text_PART.(Grid.Column)" Value="1" />
-						<Setter Target="Text_PART.(Grid.ColumnSpan)" Value="1" />
-					</VisualState.Setters>
-				</VisualState>
-
-				<VisualState x:Name="VerticalState">
-					<VisualState.Setters>
-						<Setter Target="Label_PART.(Grid.Row)" Value="0" />
-						<Setter Target="Label_PART.(Grid.RowSpan)" Value="1" />
-						<Setter Target="Label_PART.(Grid.Column)" Value="0" />
-						<Setter Target="Label_PART.(Grid.ColumnSpan)" Value="2" />
-
-						<Setter Target="TextEdit_PART.(Grid.Row)" Value="1" />
-						<Setter Target="TextEdit_PART.(Grid.RowSpan)" Value="1" />
-						<Setter Target="TextEdit_PART.(Grid.Column)" Value="0" />
-						<Setter Target="TextEdit_PART.(Grid.ColumnSpan)" Value="2" />
-
-						<Setter Target="Text_PART.(Grid.Row)" Value="1" />
-						<Setter Target="Text_PART.(Grid.RowSpan)" Value="1" />
-						<Setter Target="Text_PART.(Grid.Column)" Value="0" />
-						<Setter Target="Text_PART.(Grid.ColumnSpan)" Value="2" />
-					</VisualState.Setters>
-				</VisualState>
-			</VisualStateGroup>
-		</VisualStateManager.VisualStateGroups>
-	</Grid>
-
-</UserControl>
