+++ src/Files.App.Controls/PropertiesView/PropertiesView.ThemeResources.xaml
@@ -3,7 +3,8 @@
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:animatedvisuals="using:Microsoft.UI.Xaml.Controls.AnimatedVisuals"
-	xmlns:controls="using:Microsoft.UI.Xaml.Controls">
+	xmlns:controls="using:Microsoft.UI.Xaml.Controls"
+	xmlns:fcontrols="using:Files.App.Controls">
 
 
 	<!--  THEME RESOURCE DICTIONARY  -->
@@ -125,7 +126,7 @@
 
 
 
-	<!--  EXPANDER  -->
+	<!--  PROPERTIES EXPANDER  -->
 	<Style x:Key="PropertiesExpanderStyle" TargetType="controls:Expander">
 		<!--
 			We won't make this control tab stoppable, since the tab focus should go
@@ -306,7 +307,7 @@
 
 
 
-	<!--  EXPANDER DOWN BUTTON  -->
+	<!--  vEXPANDER DOWN BUTTON  -->
 	<Style x:Key="PropertiesExpanderHeaderDownStyle" TargetType="ToggleButton">
 		<Setter Property="Padding" Value="{ThemeResource PropertiesExpanderHeaderPadding}" />
 		<Setter Property="Background" Value="{ThemeResource PropertiesExpanderHeaderBackground}" />
@@ -465,7 +466,7 @@
 	</Style>
 
 
-	<!--  EXPANDER UP BUTTON  -->
+	<!--  PROPERTIES EXPANDER UP BUTTON  -->
 	<Style
 		x:Key="PropertiesExpanderHeaderUpStyle"
 		BasedOn="{StaticResource PropertiesExpanderHeaderDownStyle}"
@@ -621,6 +622,144 @@
 						</VisualStateManager.VisualStateGroups>
 					</Border>
 				</ControlTemplate>
+			</Setter.Value>
+		</Setter>
+	</Style>
+
+
+
+
+	<!--  PROPERTIES VIEW ITEM - TEXT  -->
+	<Style BasedOn="{StaticResource DefaultPropertiesViewItem_Text}" TargetType="fcontrols:PropertiesViewItemText" />
+
+	<Style x:Key="DefaultPropertiesViewItem_Text" TargetType="fcontrols:PropertiesViewItemText">
+
+		<Setter Property="Template">
+			<Setter.Value>
+				<ControlTemplate TargetType="fcontrols:PropertiesViewItemText">
+					<Grid
+						Padding="12,6"
+						HorizontalAlignment="Stretch"
+						VerticalAlignment="Stretch">
+						<Grid.ColumnDefinitions>
+							<ColumnDefinition Width="{TemplateBinding LabelWidth}" />
+							<ColumnDefinition Width="*" />
+						</Grid.ColumnDefinitions>
+						<Grid.RowDefinitions>
+							<RowDefinition Height="Auto" />
+							<RowDefinition Height="Auto" />
+						</Grid.RowDefinitions>
+
+						<TextBlock
+							x:Name="PART_Label"
+							Grid.Row="0"
+							Grid.RowSpan="2"
+							Grid.Column="0"
+							Grid.ColumnSpan="1"
+							Margin="0,2,12,2"
+							VerticalAlignment="Center"
+							Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+							Style="{ThemeResource CaptionTextBlockStyle}"
+							Text="{TemplateBinding Label}" />
+
+						<TextBlock
+							x:Name="PART_Text"
+							Grid.Row="0"
+							Grid.RowSpan="2"
+							Grid.Column="1"
+							Grid.ColumnSpan="1"
+							Margin="0,2,12,2"
+							VerticalAlignment="Center"
+							Text="{TemplateBinding Text}"
+							TextWrapping="WrapWholeWords"
+							Visibility="Visible" />
+
+						<TextBox
+							x:Name="PART_TextEdit"
+							Grid.Row="0"
+							Grid.RowSpan="2"
+							Grid.Column="1"
+							Grid.ColumnSpan="1"
+							Margin="-12,2,0,2"
+							VerticalAlignment="Center"
+							Text="{TemplateBinding Text}"
+							Visibility="Collapsed" />
+
+
+
+
+						<VisualStateManager.VisualStateGroups>
+							<VisualStateGroup x:Name="EditableStateGroup">
+								<VisualState x:Name="NotEditableState">
+									<VisualState.Setters>
+										<Setter Target="TextEdit_PART.Visibility" Value="Collapsed" />
+										<Setter Target="Text_PART.Visibility" Value="Visible" />
+									</VisualState.Setters>
+								</VisualState>
+								<VisualState x:Name="EditableState">
+									<VisualState.Setters>
+										<Setter Target="TextEdit_PART.Visibility" Value="Visible" />
+										<Setter Target="Text_PART.Visibility" Value="Collapsed" />
+									</VisualState.Setters>
+								</VisualState>
+							</VisualStateGroup>
+							<VisualStateGroup x:Name="OrientationStateGroup">
+								<VisualState x:Name="HorizontalState">
+									<VisualState.Setters>
+										<Setter Target="PART_Label.(Grid.Row)" Value="0" />
+										<Setter Target="PART_Label.(Grid.RowSpan)" Value="2" />
+										<Setter Target="PART_Label.(Grid.Column)" Value="0" />
+										<Setter Target="PART_Label.(Grid.ColumnSpan)" Value="1" />
+
+										<Setter Target="PART_TextEdit.(Grid.Row)" Value="0" />
+										<Setter Target="PART_TextEdit.(Grid.RowSpan)" Value="2" />
+										<Setter Target="PART_TextEdit.(Grid.Column)" Value="1" />
+										<Setter Target="PART_TextEdit.(Grid.ColumnSpan)" Value="1" />
+
+										<Setter Target="PART_Text.(Grid.Row)" Value="0" />
+										<Setter Target="PART_Text.(Grid.RowSpan)" Value="2" />
+										<Setter Target="PART_Text.(Grid.Column)" Value="1" />
+										<Setter Target="PART_Text.(Grid.ColumnSpan)" Value="1" />
+									</VisualState.Setters>
+								</VisualState>
+
+								<VisualState x:Name="VerticalState">
+									<VisualState.Setters>
+										<Setter Target="PART_Label.(Grid.Row)" Value="0" />
+										<Setter Target="PART_Label.(Grid.RowSpan)" Value="1" />
+										<Setter Target="PART_Label.(Grid.Column)" Value="0" />
+										<Setter Target="PART_Label.(Grid.ColumnSpan)" Value="2" />
+
+										<Setter Target="PART_TextEdit.(Grid.Row)" Value="1" />
+										<Setter Target="PART_TextEdit.(Grid.RowSpan)" Value="1" />
+										<Setter Target="PART_TextEdit.(Grid.Column)" Value="0" />
+										<Setter Target="PART_TextEdit.(Grid.ColumnSpan)" Value="2" />
+
+										<Setter Target="PART_Text.(Grid.Row)" Value="1" />
+										<Setter Target="PART_Text.(Grid.RowSpan)" Value="1" />
+										<Setter Target="PART_Text.(Grid.Column)" Value="0" />
+										<Setter Target="PART_Text.(Grid.ColumnSpan)" Value="2" />
+									</VisualState.Setters>
+								</VisualState>
+							</VisualStateGroup>
+						</VisualStateManager.VisualStateGroups>
+					</Grid>
+				</ControlTemplate>
+			</Setter.Value>
+		</Setter>
+	</Style>
+
+
+
+
+	<!--  PROPERTIES VIEW ITEM - RATING  -->
+	<Style BasedOn="{StaticResource DefaultPropertiesViewItem_Rating}" TargetType="fcontrols:PropertiesViewItemRating" />
+
+	<Style x:Key="DefaultPropertiesViewItem_Rating" TargetType="fcontrols:PropertiesViewItemRating">
+
+		<Setter Property="Template">
+			<Setter.Value>
+				<ControlTemplate TargetType="fcontrols:PropertiesViewItemRating" />
 			</Setter.Value>
 		</Setter>
 	</Style>
