+++ src/Files.App.Controls/Themes/Generic.xaml
@@ -80,5 +80,4 @@
 		<!--#endregion-->
 
 	</ResourceDictionary.MergedDictionaries>
-
 </ResourceDictionary>
